refactor: interface-only settings
