+++ frontend/app/src/locales/en.json
@@ -3979,7 +3979,8 @@
       },
       "select_color": "Select Color",
       "subtitle": "Manage colors for the app theme",
-      "title": "Theme"
+      "title": "Theme",
+      "text": "App theme"
     },
     "theme_switch": {
       "auto": "Auto (Detect Device Theme)",
